Make `rabbitmq-queues` commands follow standard naming pattern
