+++ lib/rabbitmq/cli/queues/commands/add_member_command.ex
@@ -13,31 +13,20 @@
 ## The Initial Developer of the Original Code is GoPivotal, Inc.
 ## Copyright (c) 2007-2019 Pivotal Software, Inc.  All rights reserved.
 
-defmodule RabbitMQ.CLI.Queues.Commands.DeleteMember do
+defmodule RabbitMQ.CLI.Queues.Commands.AddMemberCommand do
   import Rabbitmq.Atom.Coerce
 
   @behaviour RabbitMQ.CLI.CommandBehaviour
 
-  defp default_opts, do: %{vhost: "/"}
-
   def merge_defaults(args, opts) do
-    {args, Map.merge(default_opts(), opts)}
+    {args, Map.merge(%{vhost: "/"}, opts)}
   end
 
-  def validate(args, _) when length(args) < 2 do
-    {:validation_failure, :not_enough_args}
-  end
-
-  def validate(args, _) when length(args) > 2 do
-    {:validation_failure, :too_many_args}
-  end
-
-  def validate([_, _], _), do: :ok
-
+  use RabbitMQ.CLI.Core.AcceptsTwoPositionalArguments
   use RabbitMQ.CLI.Core.RequiresRabbitAppRunning
 
   def run([name, node] = _args, %{vhost: vhost, node: node_name}) do
-    case :rabbit_misc.rpc_call(node_name, :rabbit_quorum_queue, :delete_member, [
+    case :rabbit_misc.rpc_call(node_name, :rabbit_quorum_queue, :add_member, [
            vhost,
            name,
            to_atom(node)
@@ -53,8 +42,8 @@ defmodule RabbitMQ.CLI.Queues.Commands.DeleteMember do
   use RabbitMQ.CLI.DefaultOutput
 
   def banner([name, node], _) do
-    "Deleting member #{node} from quorum queue #{name} cluster..."
+    "Adding member #{node} to quorum queue #{name} cluster..."
   end
 
-  def usage, do: "delete_member [-p <vhost>] <queuename> <node>"
+  def usage, do: "add_member [-p <vhost>] <queuename> <node>"
 end
