+++ lib/rabbitmq/cli/queues/commands/shrink_command.ex
@@ -13,7 +13,7 @@
 ## The Initial Developer of the Original Code is GoPivotal, Inc.
 ## Copyright (c) 2007-2019 Pivotal Software, Inc.  All rights reserved.
 
-defmodule RabbitMQ.CLI.Queues.Commands.Shrink do
+defmodule RabbitMQ.CLI.Queues.Commands.ShrinkCommand do
   import Rabbitmq.Atom.Coerce
 
   @behaviour RabbitMQ.CLI.CommandBehaviour
